adding members to project

+++ ProjectsAndTasksModule/src/main/java/projectsandtasks/models/Member.java
@@ -9,44 +9,45 @@ import javax.persistence.*;
  * Created by bake on 3/20/17.
  */
 @Entity
-@JsonIdentityInfo(generator=ObjectIdGenerators.PropertyGenerator.class, property="id")
+@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "id")
 public class Member {
-    public Long getId() {
-        return id;
-    }
+  public Long getId() {
+    return id;
+  }
 
-    public void setId(Long id) {
-        this.id = id;
-    }
+  public void setId(Long id) {
+    this.id = id;
+  }
 
-    public int getUserId() {
-        return userId;
-    }
+  public Long getUserId() {
+    return userId;
+  }
 
-    public void setUserId(int userId) {
-        this.userId = userId;
-    }
+  public void setUserId(Long userId) {
+    this.userId = userId;
+  }
 
   public Project getProject() {
-        return project;
-    }
+    return project;
+  }
 
-    public void setProject(Project project) {
-        this.project = project;
-    }
+  public void setProject(Project project) {
+    this.project = project;
+  }
 
-    @GeneratedValue(strategy=GenerationType.AUTO)
-    @Id
-    private Long id;
-    private int userId;
-    @ManyToOne
-    @JoinColumn(name = "projectId")
-    private Project project;
+  @GeneratedValue(strategy = GenerationType.AUTO)
+  @Id
+  private Long id;
+  private Long userId;
+  @ManyToOne
+  @JoinColumn(name = "projectId")
+  private Project project;
 
-    protected Member () {    }
+  protected Member() {
+  }
 
-    public Member(int userId, Project project) {
-        this.userId = userId;
-        this.project = project;
-    }
+  public Member(Long userId, Project project) {
+    this.userId = userId;
+    this.project = project;
+  }
 }
